refactor: address review comments on request logging
- Replace SqliteRequestLogger with NoopRequestLogger as default
- Move PgRequestLogger impl to StyleAgentLogs in domain crate
- Use Box<dyn Error> instead of anyhow in RequestLogger trait
- Add results_json field to log actual search results returned
- Wrap StyleAgentLogs in Arc in App for shared ownership
- McpGateway gets logger from app.style_agent_logs() internally
- Remove stale deps from mcp-gateway (chrono, sqlx, async-trait)
