+++ domain/Cargo.toml
@@ -6,6 +6,9 @@ edition.workspace = true
 [lib]
 
 [dependencies]
+style-agent-server = { workspace = true }
+anyhow = { workspace = true }
+async-trait = "0.1"
 base64 = "0.22"
 chrono = { version = "0.4", features = ["clock", "serde"], default-features = false }
 derive_builder = "0.20"
