+++ mcp-gateway/Cargo.toml
@@ -7,16 +7,12 @@ edition.workspace = true
 
 [dependencies]
 galoy-agents-domain = { path = "../domain" }
-style-agent-core = { workspace = true }
 style-agent-server = { workspace = true }
 anyhow = { workspace = true }
-async-trait = "0.1"
 axum = { workspace = true }
 http = "1"
 rmcp = { version = "1.2", features = ["server", "transport-streamable-http-server", "macros"] }
 schemars = "0.8"
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
-chrono = { version = "0.4", default-features = false, features = ["clock", "serde"] }
-sqlx = { version = "0.8", default-features = false, features = ["runtime-tokio-rustls", "postgres", "chrono"] }
 tracing = "0.1"
